+++ dwds/test/test_context.dart
@@ -6,6 +6,7 @@ import 'package:dwds/service.dart';
 import 'package:dwds/src/chrome_proxy_service.dart';
 import 'package:dwds/src/helpers.dart';
 import 'package:http/http.dart' as http;
+import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 import 'package:webdriver/io.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
@@ -31,9 +32,11 @@ class TestContext {
   String path;
 
   TestContext(
-      {this.directory,
+      {String directory,
       this.path = 'hello_world/index.html',
-      this.pathToServe = 'example'});
+      this.pathToServe = 'example'}) {
+       this.directory = directory ?? p.relative('../_test', from: p.current);
+      }
 
   Future<void> setUp() async {
     port = await findUnusedPort();
@@ -45,16 +48,11 @@ class TestContext {
           'Could not start ChromeDriver. Is it installed?\nError: $e');
     }
 
+    await Process.run('pub', ['get'],
+        workingDirectory: directory);
+
     webdev = await Process.start(
-        'pub',
-        [
-          'run',
-          'webdev',
-          'serve',
-          '$pathToServe:$port',
-          '--',
-          '--delete-conflicting-outputs'
-        ],
+        'pub', ['run', 'webdev', 'serve', '$pathToServe:$port'],
         workingDirectory: directory);
     webdev.stderr
         .transform(const Utf8Decoder())