+++ dwds/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Move `data` abstractions from `package:webdev` into `package:dwds`.
 - Move debugging related handlers from `package:webdev` into `package:dwds`.
+- Move injected client from `package:webdev` into `package:dwds`.
 
 ## 0.3.3